added changing nickname and nick color
